requirements.txt
PyMuPDF>=1.24.0
Shapely>=2.0.0
networkx>=3.0
numpy>=1.24.0
PyYAML>=6.0
pytest>=7.0
